v4.0.0-beta config injection pattern, refactor into different task files (bug 1145421)
