+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "marketplace-gulp",
   "description": "gulpfiles for Firefox Marketplace frontend projects",
-  "version": "3.4.0",
+  "version": "4.0.0-beta",
   "main": "index",
   "repository": {
     "url": "git://github.com/mozilla/marketplace-gulp.git",
@@ -14,7 +14,6 @@
   "dependencies": {
     "commonplace": "0.6.45",
     "del": "0.1.3",
-    "event-stream": "3.1.7",
     "gulp": "3.8.x",
     "gulp-concat": "2.4.1",
     "gulp-concat-css": "1.0.0",
@@ -28,8 +27,6 @@
     "gulp-rename": "1.2.0",
     "gulp-replace": "0.4.0",
     "gulp-stylus": "2.0.1",
-    "gulp-uglify": "1.1.x",
-    "gulp-util": "3.0.x",
     "gulp-watch": "1.1.0",
     "gulp-webserver": "0.9.0",
     "gulp-zip": "2.0.2",
@@ -42,7 +39,6 @@
     "requirejs": "2.1.15",
     "run-sequence": "1.0.1",
     "stylus": "0.50.x",
-    "through2": "0.6.2",
-    "underscore": "^1.8.2"
+    "through2": "0.6.2"
   }
 }
