+++ plugins/imgurls-absolutify.js
@@ -5,8 +5,6 @@
 var through = require('through2');
 var path = require('path');
 
-var config = require(process.env.GULP_CONFIG_PATH || '../../../config');
-
 var url_pattern = /url\(([^)]+)\)/g;
 
 
@@ -20,7 +18,7 @@ function transform(file) {
         // Ensure it is an absolute URL (no relative URLs).
         var has_origin = url.search(/(https?):|\/\//) === 0 || url[0] === '/';
         if (!has_origin) {
-            var absolute_path = path.join(config.CSS_DEST_PATH, url);
+            var absolute_path = path.join(MKT_CONFIG.CSS_DEST_PATH, url);
             url = '/' + path.relative('src', absolute_path);
         }
